Merge dev -> master 2

+++ apps/billing/views/invoice.py
@@ -4,6 +4,7 @@ from itertools import chain
 
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin
+from django.core.files.base import ContentFile
 from django.db.models import DecimalField, ExpressionWrapper, F, Sum
 from django.http import HttpResponse
 from django.shortcuts import redirect, render
@@ -11,9 +12,10 @@ from django.urls import reverse_lazy
 from django.views.generic import DeleteView, DetailView, FormView, TemplateView, View
 
 from apps.activity.models import ExpenseEntry, TimeEntry
-from apps.invoicing.forms import InvoiceForm
-from apps.invoicing.functions import generate_invoice
-from apps.invoicing.models import Invoice
+from apps.billing.forms import InvoiceForm
+from apps.billing.forms.invoice import EditInvoiceForm
+from apps.billing.functions import generate_invoice
+from apps.billing.models import Invoice
 from apps.matters.models import Matter
 
 
@@ -26,33 +28,33 @@ def index(request):
     )
 
     context = {
-        "page": "invoicing",
+        "page": "billing",
         "invoices": invoices,
     }
 
-    return render(request, "invoicing/list.html", context)
+    return render(request, "billing/list.html", context)
 
 
 class InvoiceDetailView(LoginRequiredMixin, DetailView):
     model = Invoice
-    template_name = "invoicing/preview/preview.html"
+    template_name = "billing/preview/preview.html"
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
 
-        context["page"] = "invoicing"
+        context["page"] = "billing"
 
         context["file_url"] = reverse_lazy(
-            "invoicing:invoice-pdf", kwargs={"pk": self.object.pk}
+            "billing:invoice-pdf", kwargs={"pk": self.object.pk}
         )
 
         return context
 
 
 class AddInvoiceView(LoginRequiredMixin, FormView):
-    template_name = "invoicing/form-invoice.html"
+    template_name = "billing/form-invoice.html"
     form_class = InvoiceForm
-    success_url = reverse_lazy("invoicing:invoicing")
+    success_url = reverse_lazy("billing:billing")
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
@@ -106,17 +108,79 @@ class AddInvoiceView(LoginRequiredMixin, FormView):
         return super().form_valid(form)
 
 
+class EditInvoiceView(LoginRequiredMixin, FormView):
+    template_name = "billing/edit-invoice.html"
+    form_class = EditInvoiceForm
+
+    def get_success_url(self):
+        return reverse_lazy("billing:invoice-detail", kwargs={"pk": self.kwargs["pk"]})
+
+    def get_form_kwargs(self):
+        kwargs = super().get_form_kwargs()
+
+        invoice = Invoice.objects.get(pk=self.kwargs["pk"])
+        kwargs["instance"] = invoice
+
+        return kwargs
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+
+        invoice = Invoice.objects.get(pk=self.kwargs["pk"])
+        context["invoice"] = invoice
+
+        return context
+
+    def form_valid(self, form):
+        form.save()
+
+        invoice = Invoice.objects.get(pk=self.kwargs["pk"])
+
+        date_limit = form.cleaned_data["date_limit"]
+
+        # Remove entries that are not within the new date limit
+        TimeEntry.objects.filter(invoice=invoice, date__gt=date_limit).update(
+            invoice=None
+        )
+
+        ExpenseEntry.objects.filter(invoice=invoice, date__gt=date_limit).update(
+            invoice=None
+        )
+
+        time_entry_amount = (
+            TimeEntry.objects.filter(invoice=invoice)
+            .annotate(
+                fee=ExpressionWrapper(
+                    F("hours") * F("rate"), output_field=DecimalField()
+                )
+            )
+            .aggregate(total_fee=Sum("fee"))["total_fee"]
+        ) or 0
+
+        expense_amount = (
+            ExpenseEntry.objects.filter(invoice=invoice).aggregate(
+                total_amount=Sum("amount")
+            )["total_amount"]
+            or 0
+        )
+
+        invoice.amount = (time_entry_amount + expense_amount) - invoice.discount
+        invoice.save()
+
+        return super().form_valid(form)
+
+
 class DeleteInvoiceView(LoginRequiredMixin, DeleteView):
     model = Invoice
-    success_url = reverse_lazy("invoicing:invoicing")
+    success_url = reverse_lazy("billing:billing")
 
     def get(self, request, *args, **kwargs):
         return self.delete(request, *args, **kwargs)
 
 
 class CancelInvoiceView(LoginRequiredMixin, TemplateView):
-    template_name = "invoicing/confirm-cancel.html"
-    success_url = reverse_lazy("invoicing:invoicing")
+    template_name = "billing/confirm-cancel.html"
+    success_url = reverse_lazy("billing:billing")
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
@@ -133,15 +197,22 @@ class InvoicePDFView(LoginRequiredMixin, DetailView):
     def get(self, request, *args, **kwargs):
         invoice = self.get_object()
 
-        file = generate_invoice(invoice, request)
+        if invoice.status == "CANCELED":
+            with open(invoice.pdf_file.path, "rb") as pdf:
+                response = HttpResponse(pdf.read(), content_type="application/pdf")
+                response["Content-Disposition"] = (
+                    f'filename="Invoice {invoice.id} - {invoice.matter} - {invoice.date_issued}.pdf"'
+                )
+        else:
+            file = generate_invoice(invoice, request)
 
-        with open(file.name, "rb") as pdf:
-            response = HttpResponse(pdf.read(), content_type="application/pdf")
-            response["Content-Disposition"] = (
-                f'filename="Invoice {invoice.id} - {invoice.matter} - {invoice.date_issued}.pdf"'
-            )
+            with open(file.name, "rb") as pdf:
+                response = HttpResponse(pdf.read(), content_type="application/pdf")
+                response["Content-Disposition"] = (
+                    f'filename="Invoice {invoice.id} - {invoice.matter} - {invoice.date_issued}.pdf"'
+                )
 
-        os.unlink(file.name)
+            os.unlink(file.name)
 
         return response
 
@@ -166,9 +237,19 @@ class StatusUpdateView(LoginRequiredMixin, View):
         invoice.save()
 
         if invoice_status == "CANCELED":
+            pdf_file = generate_invoice(invoice, request)
+
+            with open(pdf_file.name, "rb") as pdf:
+                invoice.pdf_file.save(
+                    f"Invoice {invoice.id} - {invoice.matter} - {invoice.date_issued}.pdf",
+                    ContentFile(pdf.read()),
+                )
+
+            invoice.save()
+
             TimeEntry.objects.filter(invoice=invoice).update(invoice=None)
             ExpenseEntry.objects.filter(invoice=invoice).update(invoice=None)
 
-            return redirect("invoicing:invoicing")
+            return redirect("billing:billing")
 
-        return render(request, "invoicing/invoice-row.html", {"invoice": invoice})
+        return render(request, "billing/invoice-row.html", {"invoice": invoice})
